pyproject.toml
[project]
name = "sre-incident-commander"
version = "0.1.0"
description = "OpenEnv SRE DevOps Simulator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "numpy>=1.26.0",
    "gymnasium>=0.29.1",
    "openai>=1.14.0",
    "openenv-core>=0.1.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0"
]
[project.scripts]
server = "server.app:main"